+++ src/app/translations/translation-detail/translation-detail.component.ts
@@ -43,6 +43,12 @@ export class TranslationDetailComponent implements OnInit {
 
   // Delete translation
   deleteTranslation() {
-    this.translationSvc.deleteTranslation(this.translation.$key)
+    if(this.confirmAction('Are you sure you want to delete this phrase?'))
+      this.translationSvc.deleteTranslation(this.translation.$key)
+  }
+
+  // Default confirm action
+  private confirmAction(msg: string) {
+    return confirm(msg);
   }
 }
